Confirm delete_site_option() exists before calling it during an update.
This function won't yet exist when updating from 2.7 [sic].

Merges [25893] to the 3.7 branch.
fixes #25682.

Built from https://develop.svn.wordpress.org/branches/3.7@25894


git-svn-id: http://core.svn.wordpress.org/branches/3.7@25806 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/update-core.php
@@ -894,7 +894,8 @@ function update_core($from, $to) {
 	do_action( '_core_updated_successfully', $wp_version );
 
 	// Clear the option that blocks auto updates after failures, now that we've been successful.
-	delete_site_option( 'auto_core_update_failed' );
+	if ( function_exists( 'delete_site_option' ) )
+		delete_site_option( 'auto_core_update_failed' );
 
 	return $wp_version;
 }
